# docs/swagger.yaml
definitions:
...
  models.Blog:
    properties:
      bloggerlink:
        type: string
      bloggername:
        type: string
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      link:
        type: string
      postdate:
        type: string
      priority:
        type: integer
      query:
        type: string
      title:
        type: string
    type: object
...
  models.News:
    properties:
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      link:
        type: string
      press:
        type: string
      priority:
        type: integer
      pubDate:
        type: string
      query:
        type: string
      title:
        type: string
    type: object
...
  models.Youtube:
    properties:
      channelID:
        type: string
      channelTitle:
        type: string
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      publishedAt:
        type: string
      query:
        type: string
      thumbnailURL:
        type: string
      title:
        type: string
      videoID:
        type: string
    type: object
  v2.AdjAuctionQuantityBarGraph:
    properties:
      graphBar:
        items:
          $ref: '#/definitions/v2.adjAuctionQuantityGraphBar'
        type: array
      rangeLabel:
        items:
          type: string
        type: array
      stdItemCode:
        type: string
    type: object
...
  v2.Blog:
    properties:
      bloggerlink:
        type: string
      bloggername:
        type: string
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      link:
        type: string
      postdate:
        type: string
      priority:
        type: integer
      query:
        type: string
      title:
        type: string
    type: object
  v2.ItemCodeMap:
    properties:
      examinItemCode:
        type: string
      hskPrdlstCode:
        type: string
      stdItemCode:
        type: string
    type: object
  v2.News:
    properties:
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      link:
        type: string
      press:
        type: string
      priority:
        type: integer
      pubDate:
        type: string
      query:
        type: string
      title:
        type: string
    type: object
...
  v2.RetailPrice:
    properties:
      examinDate:
        type: string
      examinGradeCode:
        type: string
      examinGradeName:
        type: string
      examinItemCode:
        type: string
      examinItemName:
        type: string
      examinSpeciesCode:
        type: string
      examinSpeciesName:
        type: string
      examinUnit:
        type: string
      examinUnitName:
        type: string
      maxPrice:
        type: integer
      minPrice:
        type: integer
    type: object
  v2.RetailPriceGraph:
    properties:
      graphLine:
        items:
          $ref: '#/definitions/v2.retailPriceGraphLine'
        type: array
      rangeLabel:
        items:
          type: string
        type: array
      stdItemCode:
        type: string
    type: object
  v2.StdItemKeword:
    properties:
      itemCode:
        type: string
      keyword:
        type: string
    type: object
  v2.TradeInfoGraph:
    properties:
      graphLine:
        items:
          $ref: '#/definitions/v2.tradeInfoGraphLine'
        type: array
      rangeLabel:
        items:
          type: string
        type: array
      stdItemCode:
        type: string
    type: object
...
  v2.Youtube:
    properties:
      channelID:
        type: string
      channelTitle:
        type: string
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      publishedAt:
        type: string
      query:
        type: string
      thumbnailURL:
        type: string
      title:
        type: string
      videoID:
        type: string
    type: object
  v2.adjAuctionQuantityGraphBar:
    properties:
      graphData:
        items:
          type: string
        type: array
      stdItemName:
        type: string
      stdSpeciesName:
        type: string
    type: object
  v2.retailPriceGraphData:
    properties:
      x:
        type: string
      "y":
        type: integer
    type: object
  v2.retailPriceGraphLine:
    properties:
      examinGradeName:
        type: string
      examinItemName:
        type: string
      examinSpeciesName:
        type: string
      examinUnitName:
        type: string
      graphData:
        items:
          $ref: '#/definitions/v2.retailPriceGraphData'
        type: array
      priceType:
        type: string
    type: object
  v2.tradeInfoGraphData:
    properties:
      x:
        type: string
      "y":
        type: string
    type: object
  v2.tradeInfoGraphLine:
    properties:
      graphData:
        items:
          $ref: '#/definitions/v2.tradeInfoGraphData'
        type: array
      hskPrdlstCode:
        type: string
      tradeType:
        type: string
    type: object
...
info:
  contact: {}
  description: 아그리파 API
  license: {}
  title: AGRIPA API
  version: "1.0"
paths:
...
  /media/blog:
    get:
      consumes:
      - application/json
      description: 품목별 블로그 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 품목코드
        in: query
        name: itemCode
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: baseDateTime
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: countPerPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Blog'
      summary: 품목별 블로그 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - 미디어
  /media/news:
    get:
      consumes:
      - application/json
      description: 품목별 뉴스 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 품목코드
        in: query
        name: itemCode
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: baseDateTime
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: countPerPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
      summary: 품목별 뉴스 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - 미디어
  /media/youtube:
    get:
      consumes:
      - application/json
      description: 품목별 유튜브 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 품목코드
        in: query
        name: itemCode
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: baseDateTime
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: countPerPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Youtube'
      summary: 품목별 유튜브 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - 미디어
...
  /v2/auction/adj/quantity/bar-graph/{std-item-code}:
    get:
      consumes:
      - application/json
      description: 정산 경락 거래량 바 차트 조회합니다
      parameters:
      - description: 표준품목코드
        in: path
        name: std-item-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.AdjAuctionQuantityBarGraph'
      summary: 정산 경락 거래량 바 차트 조회
      tags:
      - v2 정산 경락 거래량
...
  /v2/code/item-code-map:
    get:
      consumes:
      - application/json
      description: 품목코드 맵 조회합니다
      parameters:
      - description: 표준품목코드
        in: query
        name: stdItemCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.ItemCodeMap'
      summary: 품목코드 맵 조회
      tags:
      - v2 코드
  /v2/code/std-item-codes:
    get:
      consumes:
      - application/json
      description: 표준품목코드 키워드 조회합니다
      parameters:
      - description: 표준품목코드 검색어
        in: query
        name: query
        required: true
        type: string
      - description: 'row 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.StdItemKeword'
      summary: 표준품목코드 키워드 조회
      tags:
      - v2 코드
  /v2/media/blog:
    get:
      consumes:
      - application/json
      description: 품목별 블로그 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 표준품목코드
        in: query
        name: std-item-code
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: base-date-time
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: count-per-page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Blog'
      summary: 품목별 블로그 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - v2 미디어
  /v2/media/news:
    get:
      consumes:
      - application/json
      description: 품목별 뉴스 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 표준품목코드
        in: query
        name: std-item-code
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: base-date-time
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: count-per-page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.News'
      summary: 품목별 뉴스 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - v2 미디어
  /v2/media/youtube:
    get:
      consumes:
      - application/json
      description: 품목별 유튜브 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 표준품목코드
        in: query
        name: std-item-code
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: base-date-time
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: count-per-page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Youtube'
      summary: 품목별 유튜브 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - v2 미디어
...
  /v2/retail/price/line-graph/{std-item-code}:
    get:
      consumes:
      - application/json
      description: 소매가격 꺾은선 그래프 정보 조회합니다
      parameters:
      - description: 표준품목코드
        in: path
        name: std-item-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.RetailPriceGraph'
      summary: 소매가격 꺾은선 그래프 정보 조회
      tags:
      - v2 소매
  /v2/retail/price/previous-year/{std-item-code}:
    get:
      consumes:
      - application/json
      description: 전년 소매가격 정보 조회합니다
      parameters:
      - description: 표준품목코드
        in: path
        name: std-item-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.RetailPrice'
      summary: 전년 소매가격 정보 조회
      tags:
      - v2 소매
  /v2/retail/price/recent/{std-item-code}:
    get:
      consumes:
      - application/json
      description: 최근 소매가격 정보 조회 조회합니다
      parameters:
      - description: 표준품목코드
        in: path
        name: std-item-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.RetailPrice'
      summary: 최근 소매가격 정보 조회
      tags:
      - v2 소매
  /v2/trade/exportation/line-graph/{std-item-code}:
    get:
      consumes:
      - application/json
      description: 수출 정보 조회합니다
      parameters:
      - description: 표준품목코드
        in: path
        name: std-item-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.TradeInfoGraph'
      summary: 수출 정보 조회
      tags:
      - v2 수출입
...
  /v2/trade/importation/line-graph/{std-item-code}:
    get:
      consumes:
      - application/json
      description: 수입 정보 조회합니다
      parameters:
      - description: 표준품목코드
        in: path
        name: std-item-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.TradeInfoGraph'
      summary: 수입 정보 조회
      tags:
      - v2 수출입
...
swagger: "2.0"